[Stage] Factor out RC Config lookup and creation into a trait
